+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+## [0.12.0-ext.11] - 2018-01-02
+
 - move `save-dependencies-as-components` flag from `bit import` command to be configurable in consumer bit.json
+- misc bugs for `bit remove`
 
 ## [0.12.0-ext.10] - 2017-12-27